dark_pool_fraud_detection/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
shap>=0.42.0
lime>=0.2.0.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0
networkx>=3.1
dgl>=1.1.0
optuna>=3.2.0
